functions/openfaas/github_trello/github-issues/requirements.txt
pygithub
py-trello
environs
loguru